Restructure React Components
Restructure the React components, for a better visibility were which
component is used.

Improve mobile experience, when the kubenav is deployed in a cluster and
the website is opened on a mobile device. Therefor we are using the new
useWindowWidth hook. Next to the platform were kubenav is running on we
are now also checking the window width to show/hide components.

Rework containers view in the Pod details. The modal for a container was
removed and on desktop all details are shown in a table instead of a
list.
