+++ src/components/plugins/terminal/helpers.ts
@@ -1,9 +1,9 @@
 import SockJS from 'sockjs-client';
 import { Terminal } from 'xterm';
 
-import { IContext, ITerminalContext } from '../../declarations';
-import { kubernetesRequest, kubernetesExecRequest, kubernetesLogsRequest, sshRequest } from '../../utils/api';
-import { INCLUSTER_URL, LOG_TERMINAL_OPTIONS, SHELL_TERMINAL_OPTIONS } from '../../utils/constants';
+import { IContext, ITerminalContext } from '../../../declarations';
+import { kubernetesRequest, kubernetesExecRequest, kubernetesLogsRequest, sshRequest } from '../../../utils/api';
+import { INCLUSTER_URL, LOG_TERMINAL_OPTIONS, SHELL_TERMINAL_OPTIONS } from '../../../utils/constants';
 
 export const addShell = async (
   context: IContext,
