+++ src/components/resources/cluster/nodes/NodeDetails.tsx
@@ -7,9 +7,9 @@ import { IContext, INodeMetrics } from '../../../../declarations';
 import { kubernetesRequest } from '../../../../utils/api';
 import { AppContext } from '../../../../utils/context';
 import { formatResourceValue } from '../../../../utils/helpers';
+import AddSSH from '../../../plugins/terminal/AddSSH';
 import IonCardEqualHeight from '../../../misc/IonCardEqualHeight';
-import AddSSH from '../../../terminal/AddSSH';
-import List from '../../misc/List';
+import List from '../../misc/list/List';
 import Conditions from '../../misc/template/Conditions';
 import Configuration from '../../misc/template/Configuration';
 import Metadata from '../../misc/template/Metadata';
