+++ src/components/resources/discoveryAndLoadbalancing/ingresses/IngressDetails.tsx
@@ -14,7 +14,7 @@ import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
 import IonCardEqualHeight from '../../../misc/IonCardEqualHeight';
-import List from '../../misc/List';
+import List from '../../misc/list/List';
 import Configuration from '../../misc/template/Configuration';
 import Metadata from '../../misc/template/Metadata';
 import Row from '../../misc/template/Row';
