+++ src/components/resources/misc/details/DeleteItem.tsx
