+++ src/components/resources/misc/details/EditItem.tsx
