+++ src/components/resources/misc/details/RestartItem.tsx
