+++ src/components/resources/misc/details/ScaleItem.tsx
