+++ src/components/resources/workloads/daemonSets/DaemonSetDetails.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
 import { labelSelector } from '../../../../utils/helpers';
-import List from '../../misc/List';
+import List from '../../misc/list/List';
 import Configuration from '../../misc/template/Configuration';
 import Metadata from '../../misc/template/Metadata';
 import Row from '../../misc/template/Row';
