+++ src/components/resources/workloads/jobs/JobDetails.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
 import { labelSelector, timeDifference } from '../../../../utils/helpers';
-import List from '../../misc/List';
+import List from '../../misc/list/List';
 import Conditions from '../../misc/template/Conditions';
 import Configuration from '../../misc/template/Configuration';
 import Metadata from '../../misc/template/Metadata';
